+++ RickAndMorty/DataFetcher/RMDataProvider.swift
@@ -45,7 +45,6 @@ struct RMEntity<T: Decodable, C: Configuration> {
     func endpointForPage(page: Int) -> RMEndpoint {
         configuration.endpointForPage(page: page)
     }
-    
 }
 
 struct RMDataProvider<A: Decodable, B: Configuration> {
@@ -61,55 +60,4 @@ struct RMDataProvider<A: Decodable, B: Configuration> {
     func getPage(page: Int) async throws -> RMResponse<A> {
         try await networkFetcher.fetch(endpoint: entity.endpointForPage(page: page))
     }
-}
-
-struct RMCharacterConfiguration: Configuration {
-    
-    var cellHorizontalType: RMItemCell.Type {
-        RMCharacterHorizontalCell.self
-    }
-    
-    var reuseID: String {
-        "RMCharacterHorizontalCell"
-    }
-    
-    func endpointForPage(page: Int) -> RMEndpoint {
-        RMCharacterEndpoint.characterPage(page)
-    }
-    
-    func getCellViewModelForEntity<T: Decodable>(entity: T) -> RMItemCellViewModel? {
-        guard let rmCharacter = entity as? RMCharacter else {
-            return nil
-        }
-        
-        let viewModel = RMCharacterCellViewModel(character: rmCharacter)
-        return viewModel
-    }
-}
-
-struct RMLocationConfiguration: Configuration {
-    
-    var cellHorizontalType: RMItemCell.Type {
-        RMLocationHorizontalCell.self
-    }
-    
-    var reuseID: String {
-        "RMLocationHorizontalCell"
-    }
-    
-    func endpointForPage(page: Int) -> RMEndpoint {
-        RMLocationEndpoint.locationPage(page)
-    }
-    
-    func getCellViewModelForEntity<T>(entity: T) -> RMItemCellViewModel? where T : Decodable {
-        guard let rmLocation = entity as? RMLocation else {
-            return nil
-        }
-        
-        let viewModel = RMLocationCellViewModel(location: rmLocation)
-        return viewModel
-    }
-    
-    
-    
 }
